We created a class called Neuron that has a constructor.
The constructor sets up the neuron by:

Checking if the number of inputs is a valid number (positive integer).

Giving the neuron random weights and setting the bias and activated output to 0.

The neuron can then use these weights, bias, and inputs to make decisions later on (when you add more methods).